# internal/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "https://github.com/wetor/AnimeGo",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://www.mit-license.org/"
        }
    },
    "paths": {
...
        "/api/plugin/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "从AnimeGo中获取当前插件的配置\n插件名为不包含 'plugin' 的路径\n插件名可以忽略'.js'后缀；插件名也可以使用上层文件夹名，会自动寻找文件夹内部的 'main.js' 或 'plugin.js'\n如传入 'test'，会依次尝试寻找 'plugin/test/main.js', 'plugin/test/plugin.js', 'plugin/test.js'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "获取插件配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "插件信息",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "将当前插件的配置发送给AnimeGo并保存\n插件名为不包含 'plugin' 的路径\n插件名可以忽略'.py'后缀；插件名也可以使用上层文件夹名，会自动加载文件夹内部的 'main.py'\n如设置为 'plugin/test'，会依次尝试加载 'plugin/test/main.py', 'plugin/test.py'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "发送插件配置",
                "parameters": [
                    {
                        "description": "插件信息，data为base64编码后的json文本",
                        "name": "plugin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PluginConfigUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/plugin/manager/dir": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取插件文件夹中指定文件夹的文件列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "获取插件文件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.DirResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建插件文件夹中的文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "创建文件夹",
                "parameters": [
                    {
                        "description": "创建文件夹信息",
                        "name": "path",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PathRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除插件文件夹中指定文件文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "删除文件夹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/plugin/manager/file": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取插件文件夹中指定文件的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "获取插件文件内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建或修改插件文件夹中指定文件",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "创建或修改插件文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "文件内容",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除插件文件夹中指定插件文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "删除插件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/plugin/manager/rename": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "重命名插件文件夹中的文件或文件夹",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugin"
                ],
                "summary": "重命名文件或文件夹",
                "parameters": [
                    {
                        "description": "重命名信息",
                        "name": "path",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewPathRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "300": {
                        "description": "Multiple Choices",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DirResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "models.File": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "delete": {
                    "type": "boolean"
                },
                "is_dir": {
                    "type": "boolean"
                },
                "modify_time": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "read": {
                    "type": "boolean"
                },
                "size": {
                    "type": "integer"
                },
                "write": {
                    "type": "boolean"
                }
            }
        },
        "models.NewPathRequest": {
            "type": "object",
            "properties": {
                "new_path": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "models.PathRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "models.PluginConfigUploadRequest": {
            "type": "object",
            "required": [
                "data",
                "name"
            ],
            "properties": {
                "data": {
                    "description": "base64格式的文本数据",
                    "type": "string"
                },
                "name": {
                    "description": "插件文件名",
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Access-Key",
            "in": "header"
        }
    }
}